pyproject.toml
[project]
name = "stock-watchdog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "apscheduler>=3.11.0",
    "asyncio-mqtt>=0.16.2",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "numpy>=2.2.6",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "uvicorn[standard]>=0.34.3",
    "yfinance>=0.2.61",
]
